requirements.txt
numpy
Pillow
scikit-image
scipy
matplotlib
gradio
scikit-learn
joblib
opencv-python
torch
torchvision